Treep
I volunteered at San Francisco Botanical Garden when I was a student at
Academy of Art. I took groups of children to find and learn about
Redwood trees. Other than that, there are over 8,000 trees in the
garden that can be explored. The garden is a like a "zoo" of trees.

As a interactive designer, I created the app Treep, which can be a
great in-hand guide for volunteers or parents who lead children to the
garden. This app also became my final thesis project at school. I
finished the app on my own from conceptual develoment, user testing, UI
design, to coding.